pyproject.toml
[project]
name = "walletservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "faker>=38.2.0",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
